app/test/pyproject.toml
[tool.poetry]
description = "Lambda Functions for automate rule management in AWS Network Firewall"
name = "app"
version = "0.1.13"

[tool.poetry.dependencies]
aws-lambda-powertools = "^2.25.1"
aws-xray-sdk = "^2.12.0"
bandit = "^1.7.5"
jsonschema = "^4.19.1"
pip-audit = "^2.6.1"
pip-licenses = "^4.3.3"
python = "^3.11"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pip-licenses = "^4.3.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
